First import script fixed
Modified the import script to select against the vocabulary table.  If
an element is not in the vocab table, then it is added to the table.
This script must be run sequentially (not in parallel) because of the
interconnectedness among entries (and possible duplicate vocabulary
inserts across processes).

Also, fixed a date bug that did not handle dates that were BC/BCE.