Versioning
Versioning of the tables set up in the schema file.  Note: This
versioning system creates a 2-field primary key for each table: (id,
version).  The row for any id with the highest version number is considered the
valid version of the rows for that id.  The version number matches the
record in the version_history table for who edited that entry.  Version
numbers are not matched across tables.

So, for example, to find all the names for a cpf record, one would get
the cpf_id in question (which has it's own latest version number and the
valid flag must be true), then find all names that match the cpf_id with
the highest version number for each name id. This can be done by
searching the name view for cpf_id.